+++ skfda/exploratory/visualization/__init__.py
@@ -1,4 +1,5 @@
 from . import clustering, representation
 from ._boxplot import Boxplot, SurfaceBoxplot
 from ._magnitude_shape_plot import MagnitudeShapePlot
+from ._ddplot import DDPlot
 from .fpca import plot_fpca_perturbation_graphs
